refactor: extract ImageWithFallback and decouple Card from routing
