+++ src/features/artworks-details/index.tsx
@@ -7,18 +7,13 @@ import {
   useNavigation,
   useSearchParams,
 } from 'react-router-dom';
-import { useEffect, useState } from 'react';
+import { ImageWithFallback } from '../ui/image-with-fallback';
 
 export function ArtworksDetails() {
   const artwork = useLoaderData() as Artwork;
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const navigation = useNavigation();
-  const [imageError, setImageError] = useState(false);
-
-  useEffect(() => {
-    setImageError(false);
-  }, [artwork.id]);
 
   const isLoading =
     navigation.state === 'loading' &&
@@ -36,12 +31,7 @@ export function ArtworksDetails() {
     );
   }
 
-  const handleImageError = () => {
-    setImageError(true);
-  };
-
   const imageUrl = getArtworkImageUrl(artwork.image_id);
-  const shouldShowPlaceholder = imageError || !imageUrl;
 
   return (
     <aside className="sticky top-28 flex flex-col gap-4 rounded-lg bg-white p-4 shadow-lg ring-1 ring-gray-200">
@@ -54,18 +44,11 @@ export function ArtworksDetails() {
       </button>
 
       <div className="aspect-square w-full">
-        {shouldShowPlaceholder ? (
-          <div className="flex h-full w-full items-center justify-center rounded bg-gray-100 text-sm text-gray-400">
-            No Image Available
-          </div>
-        ) : (
-          <img
-            src={imageUrl}
-            alt={artwork.title}
-            className="h-full w-full rounded object-cover"
-            onError={handleImageError}
-          />
-        )}
+        <ImageWithFallback
+          src={imageUrl}
+          alt={artwork.title}
+          className="h-full w-full rounded object-cover"
+        />
       </div>
 
       <div className="flex flex-col">
